refactor: :recycle: reorganize imports for better structure; update data retrieval paths for consistency
