+++ frontend/app/(protected)/(admin)/gamemaster/resurrect/page.tsx
@@ -1,5 +1,5 @@
 import MainContainer from "@/components/MainContainer";
-import { getAllDeadUsers } from "@/data/admin";
+import { getAllDeadUsers } from "@/data/admin/adminUserInteractions";
 import { Button, Typography } from "@mui/material";
 import React from "react";
 import DeathCard from "./_components/DeathCard";
