+++ frontend/app/(protected)/abilities/page.tsx
@@ -3,7 +3,7 @@ import { Typography } from "@mui/material";
 import React from "react";
 import {
   getAbilityHierarchy,
-  getOwnedAbilities,
+  getUserAbilities,
 } from "@/data/abilities/getters/getAbilities";
 import { auth } from "@/auth";
 import { notFound } from "next/navigation";
@@ -17,7 +17,7 @@ export default async function AbilitiesPage() {
     return notFound();
   }
 
-  const userAbilities = await getOwnedAbilities(user?.user?.id);
+  const userAbilities = await getUserAbilities(user?.user?.id);
 
   return (
     <MainContainer>
